Option -f to force removal of existing destinations.

+++ link_everything.sh
@@ -3,9 +3,12 @@
 targets=( ".screenrc" "scripts" ".vim" ".vimrc" ".xmonad" ".zshrc" )
 for target in "${targets[@]}"
 do
-	if [ -e $target ]; then
+	if [ -e ~/$target -a "$1" = "-f" ]; then
 		echo "Removing $target"
 		rm -fR ~/$target
+	elif [ -e ~/$target ]; then
+		echo "File $target exists, not linking."
+		continue
 	fi
 	echo "Creating symlink for $target"
 	ln -s `pwd`/$target  ~/
